# contracts/openapi/openapi.yaml
openapi: 3.0.1
info:
  title: J-Bank API
  description: |
    ## 사용 전에 알아두면 좋은 것

    **1. 응답은 항상 같은 모양입니다.**
    ```json
    { "success": true, "data": { ... }, "error": null }
    { "success": false, "data": null, "error": { "code": "ACC_001_...", "message": "..." } }
    ```
    `success`만 보고 성공/실패를 먼저 판단하고, 실패면 `error.code`로 분기하세요.

    **2. 인증은 아직 없습니다(W3에서 추가 예정).**
    소유자 확인이 필요한 API(계좌 조회/상태변경/해지)는 임시로
    `customerId`를 쿼리 파라미터로 직접 받습니다. 로그인 붙기 전까지는
    프론트에서 이 값을 그대로 넘겨주세요.

    **3. 금액은 문자열입니다.**
    부동소수점 오차를 피하려고 `"10000.00"`처럼 문자열로 내려갑니다.
    숫자로 바로 파싱하지 말고 표시용 포맷 함수를 하나 두고 거기서만 다루세요.

    **4. 날짜/시각은 ISO 8601입니다.**
    예: `2026-07-21T09:30:00+09:00`.

    **5. 목록 조회는 페이지 응답입니다.**
    ```json
    { "content": [...], "page": 0, "size": 20, "totalElements": 3, "totalPages": 1 }
    ```
  version: v0 (W1)
servers:
- url: http://localhost:8080
  description: Generated server url
tags:
- name: 고객
  description: 고객 등록(CDD)과 강화된 고객확인(EDD) API
- name: 계좌
  description: 계좌 개설·조회·상태변경·해지 API
paths:
  /api/v1/customers:
    post:
      tags:
      - 고객
      summary: 고객 등록
      description: |
        신규 고객을 등록하고 CDD(고객확인)를 함께 수행합니다. 인증이 필요 없습니다.
        같은 실명번호로 다시 등록하면 409(ACC_001)가 납니다.
        응답의 eddRequired가 true면 계좌 개설 전에 /edd API를 먼저 호출해야 합니다.
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRegisterRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseCustomerRegisterResponse"
  /api/v1/customers/{customerId}/edd:
    post:
      tags:
      - 고객
      summary: 강화된 고객확인(EDD) 등록
      description: |
        고위험으로 분류된 고객만 호출할 수 있습니다. 고위험이 아닌 고객이면
        409(ACC_004)가 납니다. 정상 처리되면 계좌 개설 게이트가 풀립니다.
      operationId: registerEdd
      parameters:
      - name: customerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EddRegisterRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseEddRegisterResponse"
  /api/v1/accounts:
    post:
      tags:
      - 계좌
      summary: 계좌 개설
      description: |
        고객의 CDD 완료 여부와 상태를 확인한 뒤 계좌를 개설합니다.
        W1 기준 초기 입금은 0원만 지원합니다(그 외 금액은 400).
      operationId: open
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountOpenRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseAccountOpenResponse"
  /api/v1/accounts/{accountId}/status:
    patch:
      tags:
      - 계좌
      summary: 계좌 상태변경
      description: |
        ACTIVE↔SUSPENDED, ACTIVE/SUSPENDED→DORMANT, DORMANT→ACTIVE만 허용합니다.
        CLOSED로의 전환은 이 API로 할 수 없습니다(해지 API를 쓰세요). 허용되지 않는
        전이는 409(ACC_007)가 납니다.
      operationId: changeStatus
      parameters:
      - name: accountId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountStatusChangeRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseAccountStatusChangeResponse"
  /api/v1/customers/{customerId}/accounts:
    get:
      tags:
      - 계좌
      summary: 고객별 계좌 목록 조회
      description: |
        한 고객이 보유한 계좌를 페이지 단위로 조회합니다. status 쿼리 파라미터로
        특정 상태(ACTIVE, SUSPENDED 등)만 필터링할 수 있습니다. 원장이 아직 없어
        balance·holdAmount·availableBalance는 항상 0입니다(W2에서 실제 값 반영).
      operationId: list
      parameters:
      - name: customerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: status
        in: query
        description: "필터링할 계좌 상태, 생략하면 전체 조회"
        required: false
        schema:
          type: string
          enum:
          - ACTIVE
          - SUSPENDED
          - DORMANT
          - CLOSED
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseCustomerAccountSummaryResponse"
  /api/v1/accounts/{accountId}:
    get:
      tags:
      - 계좌
      summary: 계좌 상세조회
      description: 인증이 아직 없어 소유자 확인용 customerId를 쿼리 파라미터로 함께 보내야 합니다. 소유자가 아니면 403.
      operationId: getDetail
      parameters:
      - name: accountId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: customerId
        in: query
        description: 임시 소유자 확인용 (W3에서 세션으로 대체 예정)
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseAccountDetailResponse"
    delete:
      tags:
      - 계좌
      summary: 계좌 해지
      description: 잔액과 지급정지 금액이 모두 0원이어야 해지할 수 있습니다. 소유자 확인용 customerId가 필요합니다.
      operationId: close
      parameters:
      - name: accountId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: customerId
        in: query
        description: 임시 소유자 확인용 (W3에서 세션으로 대체 예정)
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseAccountCloseResponse"
components:
  schemas:
    CustomerRegisterRequest:
      required:
      - birthDate
      - identityVerificationMethod
      - name
      - phone
      - residentRegNo
      type: object
      properties:
        name:
          type: string
        residentRegNo:
          type: string
        birthDate:
          type: string
          format: date
        phone:
          type: string
        address:
          type: string
        occupation:
          type: string
        identityVerificationMethod:
          type: string
          enum:
          - FACE_TO_FACE
          - NON_FACE_TO_FACE
        transactionPurpose:
          type: string
        fundSource:
          type: string
    ApiResponseCustomerRegisterResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/CustomerRegisterResponse"
        error:
          $ref: "#/components/schemas/ErrorDetail"
    CustomerRegisterResponse:
      type: object
      properties:
        customerId:
          type: string
        kycGrade:
          type: string
          enum:
          - GENERAL
          - CDD
          - EDD
        amlRiskLevel:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
        status:
          type: string
          enum:
          - ACTIVE
          - DORMANT
          - CLOSED
        eddRequired:
          type: boolean
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    EddRegisterRequest:
      required:
      - fundSource
      - supportingDocumentRef
      - transactionPurpose
      type: object
      properties:
        transactionPurpose:
          type: string
        fundSource:
          type: string
        supportingDocumentRef:
          type: string
    ApiResponseEddRegisterResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/EddRegisterResponse"
        error:
          $ref: "#/components/schemas/ErrorDetail"
    EddRegisterResponse:
      type: object
      properties:
        customerId:
          type: string
        amlRiskLevel:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
        eddCompletedAt:
          type: string
          format: date-time
    AccountOpenRequest:
      required:
      - customerId
      - initialDeposit
      - productType
      type: object
      properties:
        customerId:
          type: string
        productType:
          type: string
          enum:
          - CHECKING
        initialDeposit:
          type: number
    AccountOpenResponse:
      type: object
      properties:
        accountId:
          type: string
        accountNumber:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - SUSPENDED
          - DORMANT
          - CLOSED
        openedAt:
          type: string
          format: date-time
    ApiResponseAccountOpenResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/AccountOpenResponse"
        error:
          $ref: "#/components/schemas/ErrorDetail"
    AccountStatusChangeRequest:
      required:
      - targetStatus
      type: object
      properties:
        targetStatus:
          type: string
          enum:
          - ACTIVE
          - SUSPENDED
          - DORMANT
          - CLOSED
        reason:
          type: string
    AccountStatusChangeResponse:
      type: object
      properties:
        accountId:
          type: string
        previousStatus:
          type: string
          enum:
          - ACTIVE
          - SUSPENDED
          - DORMANT
          - CLOSED
        status:
          type: string
          enum:
          - ACTIVE
          - SUSPENDED
          - DORMANT
          - CLOSED
    ApiResponseAccountStatusChangeResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/AccountStatusChangeResponse"
        error:
          $ref: "#/components/schemas/ErrorDetail"
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    ApiResponsePageResponseCustomerAccountSummaryResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/PageResponseCustomerAccountSummaryResponse"
        error:
          $ref: "#/components/schemas/ErrorDetail"
    CustomerAccountSummaryResponse:
      type: object
      properties:
        accountId:
          type: string
        accountNumber:
          type: string
        customerId:
          type: string
        productType:
          type: string
          enum:
          - CHECKING
        status:
          type: string
          enum:
          - ACTIVE
          - SUSPENDED
          - DORMANT
          - CLOSED
        openedAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
        balance:
          type: number
        holdAmount:
          type: number
        availableBalance:
          type: number
        asOf:
          type: string
          format: date-time
    PageResponseCustomerAccountSummaryResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/CustomerAccountSummaryResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    AccountDetailResponse:
      type: object
      properties:
        accountId:
          type: string
        accountNumber:
          type: string
        customerId:
          type: string
        productType:
          type: string
          enum:
          - CHECKING
        status:
          type: string
          enum:
          - ACTIVE
          - SUSPENDED
          - DORMANT
          - CLOSED
        openedAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
    ApiResponseAccountDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/AccountDetailResponse"
        error:
          $ref: "#/components/schemas/ErrorDetail"
    AccountCloseResponse:
      type: object
      properties:
        accountId:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - SUSPENDED
          - DORMANT
          - CLOSED
        closedAt:
          type: string
          format: date-time
    ApiResponseAccountCloseResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/AccountCloseResponse"
        error:
          $ref: "#/components/schemas/ErrorDetail"